requirements.txt
ccxt
pandas
numpy
matplotlib
tensorflow
scikit-learn
binance-connector
openai